license-server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pyjwt>=2.10.0
pydantic>=2.9.0

